Bulk actions- update each action to render its own modal
